+++ index.html
@@ -12,7 +12,6 @@
     <div id="intelligenceModal"></div>
     <div id="intelligenceSettingsModal"></div>
     <div id="intelligenceGameOverModal"></div>
-    <div id="intelligenceLanguageModal"></div>
     <script type="module" src="/src/main.js"></script>
   </body>
 </html>
