+++ src/IntelligenceComponents/IntelligenceStart.vue
@@ -1,5 +1,4 @@
 <script setup lang="ts">
-import { useStorage } from "@vueuse/core";
 import { useRouter } from "vue-router";
 const router = useRouter()
