+++ src/components/MyResume.vue
@@ -8,19 +8,16 @@ const careerStore = useCareerStore()
 <template>
   <div
     class="observed-sections target mt-10 flex h-full p-2 mb-10 md:p-0 flex-col justify-center w-full items-center relative sm:h-full gap-5 transition-all duration-200">
-
     <div class="flex flex-col md:mt-0 w-full justify-center items-center">
       <h1 class="text-white font-bold text-[20px] md:text-[30px]" :style="{color : darkmode ? 'black' : ''}">{{ $t('myresume') }}</h1>
     </div>
-
     <div class="flex gap-2">
       <div class="w-4 h-4 rounded-full bg-gray-500"></div>
       <div class="w-4 h-4 rounded-full bg-gray-500"></div>
       <div class="w-4 h-4 rounded-full bg-gray-500"></div>
       <div class="w-4 h-4 rounded-full bg-gray-500"></div>
       <div class="w-4 h-4 rounded-full bg-yellow-400" :style="{backgroundColor : darkmode ? 'red' : ''}"></div>
-    </div>
-    
+    </div>  
     <div class="w-full md:w-4/5 p-1 md:p-5 gap-5 grid h-full grid-flow-row">
       <div class="grid grid-cols-1 md:grid-cols-2 gap-3">
         <div v-for="education in careerStore.careerArray">
