+++ src/components/Navbar.vue
@@ -4,7 +4,7 @@ import { onMounted, ref } from 'vue';
 import { Icon } from '@iconify/vue';
 import { useNavbarStore } from '../Store/Navbar.store';
 import i18n from '../i18n';
-
+import VNavIcon from './V-Base/V-NavIcon.vue';
 
 const navbarStore = useNavbarStore()
 const darkmode = useStorage("darkmode", false)
@@ -30,19 +30,7 @@ function changeLanguage(lang: string) {
     class="flex w-full z-[100] min-h-[35px] items-center border-b-[1px] bg-gray-800 lg:px-40 md:z-40 justify-end text-black font-bold"
     :style="[{ backgroundColor: darkmode ? '#f9f8f8' : '' }, { borderColor: darkmode ? 'red' : '' }]">
     <div class="hidden pb-2 pt-2 md:flex md:justify-between md:w-full md:items-center relative">
-
-      <div class="flex flex-col justify-center items-center">
-        <div class="w-[160px] flex justify-end">
-          <div class="border-r-[2px] border-t-[2px] w-5 h-5 border-yellow-400"
-            :style="{ borderColor: darkmode ? 'red' : '' }"></div>
-        </div>
-        <h1 class="text-white font-bold text-[20px]" :style="{ color: darkmode ? 'black' : '' }">Volkan Filazi</h1>
-        <div class="w-[160px] flex justify-start">
-          <div class="border-l-[2px] border-b-[2px] w-5 h-5 border-yellow-400"
-            :style="{ borderColor: darkmode ? 'red' : '' }"></div>
-        </div>
-      </div>
-
+      <VNavIcon>Volkan Filazi</VNavIcon>
       <ul class="flex items-center gap-3 text-white" :style="{ color: darkmode ? 'black' : '' }">
         <div v-for="navbar in navbarStore.navbarElements">
           <a class="scroll-smooth cursor-pointer border-[1px] border-transparent hover:border-white hover:transition-all hover:duration-300 p-1"
