+++ src/pages/IntelligenceSquare/IntelligenceCategory.vue
@@ -1,6 +1,6 @@
 <script setup lang="ts">
-import intelligenceNavbar from '../IntelligenceComponents/IntelligenceNavbar.vue'
-import category from '../IntelligenceComponents/Category.vue';
+import intelligenceNavbar from '../../IntelligenceComponents/IntelligenceNavbar.vue'
+import category from '../../IntelligenceComponents/Category.vue';
 import { useStorage } from '@vueuse/core';
 const darkModeOn = useStorage('darkmode',Boolean)
