+++ src/pages/Shop/ShopPortfolio.vue
@@ -1,9 +1,9 @@
 <script setup lang="ts">
 import { onMounted } from 'vue';
-import AllProducts from '../restaurantComponents/AllProducts.vue';
-import Navbar from '../restaurantComponents/NavBar.vue';
-import Drawer from '../restaurantComponents/Drawer.vue';
-import Footer from '../restaurantComponents/Footer.vue';
+import AllProducts from '../../restaurantComponents/AllProducts.vue';
+import Navbar from '../../restaurantComponents/NavBar.vue';
+import Drawer from '../../restaurantComponents/Drawer.vue';
+import Footer from '../../restaurantComponents/Footer.vue';
 
 onMounted(() =>{
   window.scrollTo(0, 0);
